feat(env-helpers): updating additional resource and session management functions
